+++ hexstrike_mcp.py
@@ -52,6 +52,7 @@ from mcp_tools.web_scan.nikto import register_nikto_tool
 from mcp_tools.web_scan.sqlmap import register_sqlmap_tool
 from mcp_tools.web_scan.wpscan import register_wpscan_tool
 from mcp_tools.web_scan.jaeles import register_jaeles_tool
+from mcp_tools.web_scan.dalfox import register_dalfox_tool
 
 from mcp_tools.exploit_framework.metasploit import register_metasploit_tool
 
@@ -72,6 +73,8 @@ from mcp_tools.url_recon.waybackurls import register_waybackurls_tool
 from mcp_tools.param_discovery.arjun import register_arjun_tool
 from mcp_tools.param_discovery.paramspider import register_paramspider_tool
 from mcp_tools.param_discovery.x8 import register_x8_tool
+
+from mcp_tools.web_probe.httpx import register_httpx_tool
 
 # Backward compatibility alias
 Colors = HexStrikeColors
@@ -286,6 +289,12 @@ TOOL_CATEGORIES = {
         lambda mcp, client, logger: register_sqlmap_tool(mcp, client, logger),
         lambda mcp, client, logger: register_wpscan_tool(mcp, client, logger),
         lambda mcp, client, logger: register_jaeles_tool(mcp, client, logger),
+        lambda mcp, client, logger: register_dalfox_tool(mcp, client, logger),
+    ],
+
+    #Tools for web probing and technology detection (e.g., httpx).
+    "web_probe": [
+        lambda mcp, client, logger: register_httpx_tool(mcp, client, logger),
     ],
 
     #Tools for automated exploitation and attack frameworks (e.g., Metasploit).
