+++ mcp_tools/web_app_security/enhanced_web_app_security.py
@@ -4,82 +4,6 @@ from typing import Dict, Any, Optional
 
 def register_enhanced_web_app_security_tools(mcp, hexstrike_client, logger, HexStrikeColors):
 
-    @mcp.tool()
-    def dalfox_xss_scan(url: str, pipe_mode: bool = False, blind: bool = False,
-                       mining_dom: bool = True, mining_dict: bool = True,
-                       custom_payload: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Dalfox for advanced XSS vulnerability scanning with enhanced logging.
-
-        Args:
-            url: The target URL
-            pipe_mode: Use pipe mode for input
-            blind: Enable blind XSS testing
-            mining_dom: Enable DOM mining
-            mining_dict: Enable dictionary mining
-            custom_payload: Custom XSS payload
-            additional_args: Additional Dalfox arguments
-
-        Returns:
-            Advanced XSS vulnerability scanning results
-        """
-        data = {
-            "url": url,
-            "pipe_mode": pipe_mode,
-            "blind": blind,
-            "mining_dom": mining_dom,
-            "mining_dict": mining_dict,
-            "custom_payload": custom_payload,
-            "additional_args": additional_args
-        }
-        logger.info(f"🎯 Starting Dalfox XSS scan: {url if url else 'pipe mode'}")
-        result = hexstrike_client.safe_post("api/tools/dalfox", data)
-        if result.get("success"):
-            logger.info(f"✅ Dalfox XSS scan completed")
-        else:
-            logger.error(f"❌ Dalfox XSS scan failed")
-        return result
-
-    @mcp.tool()
-    def httpx_probe(target: str, probe: bool = True, tech_detect: bool = False,
-                   status_code: bool = False, content_length: bool = False,
-                   title: bool = False, web_server: bool = False, threads: int = 50,
-                   additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute httpx for fast HTTP probing and technology detection.
-
-        Args:
-            target: Target file or single URL
-            probe: Enable probing
-            tech_detect: Enable technology detection
-            status_code: Show status codes
-            content_length: Show content length
-            title: Show page titles
-            web_server: Show web server
-            threads: Number of threads
-            additional_args: Additional httpx arguments
-
-        Returns:
-            Fast HTTP probing results with technology detection
-        """
-        data = {
-            "target": target,
-            "probe": probe,
-            "tech_detect": tech_detect,
-            "status_code": status_code,
-            "content_length": content_length,
-            "title": title,
-            "web_server": web_server,
-            "threads": threads,
-            "additional_args": additional_args
-        }
-        logger.info(f"🌍 Starting httpx probe: {target}")
-        result = hexstrike_client.safe_post("api/tools/httpx", data)
-        if result.get("success"):
-            logger.info(f"✅ httpx probe completed for {target}")
-        else:
-            logger.error(f"❌ httpx probe failed for {target}")
-        return result
 
     @mcp.tool()
     def anew_data_processing(input_data: str, output_file: str = "",
