+++ mcp_tools/web_probe/httpx.py
@@ -1,0 +1,45 @@
+# mcp_tools/web_probe/httpx.py
+
+from typing import Dict, Any
+
+def register_httpx_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def httpx_probe(target: str, probe: bool = True, tech_detect: bool = False,
+                   status_code: bool = False, content_length: bool = False,
+                   title: bool = False, web_server: bool = False, threads: int = 50,
+                   additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute httpx for fast HTTP probing and technology detection.
+
+        Args:
+            target: Target file or single URL
+            probe: Enable probing
+            tech_detect: Enable technology detection
+            status_code: Show status codes
+            content_length: Show content length
+            title: Show page titles
+            web_server: Show web server
+            threads: Number of threads
+            additional_args: Additional httpx arguments
+
+        Returns:
+            Fast HTTP probing results with technology detection
+        """
+        data = {
+            "target": target,
+            "probe": probe,
+            "tech_detect": tech_detect,
+            "status_code": status_code,
+            "content_length": content_length,
+            "title": title,
+            "web_server": web_server,
+            "threads": threads,
+            "additional_args": additional_args
+        }
+        logger.info(f"🌍 Starting httpx probe: {target}")
+        result = hexstrike_client.safe_post("api/tools/httpx", data)
+        if result.get("success"):
+            logger.info(f"✅ httpx probe completed for {target}")
+        else:
+            logger.error(f"❌ httpx probe failed for {target}")
+        return result
