+++ mcp_tools/web_scan/dalfox.py
@@ -1,0 +1,41 @@
+# mcp_tools/web_scan/dalfox.py
+
+from typing import Dict, Any
+
+def register_dalfox_tool(mcp, hexstrike_client, logger):
+    
+    @mcp.tool()
+    def dalfox_xss_scan(url: str, pipe_mode: bool = False, blind: bool = False,
+                       mining_dom: bool = True, mining_dict: bool = True,
+                       custom_payload: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Dalfox for advanced XSS vulnerability scanning with enhanced logging.
+
+        Args:
+            url: The target URL
+            pipe_mode: Use pipe mode for input
+            blind: Enable blind XSS testing
+            mining_dom: Enable DOM mining
+            mining_dict: Enable dictionary mining
+            custom_payload: Custom XSS payload
+            additional_args: Additional Dalfox arguments
+
+        Returns:
+            Advanced XSS vulnerability scanning results
+        """
+        data = {
+            "url": url,
+            "pipe_mode": pipe_mode,
+            "blind": blind,
+            "mining_dom": mining_dom,
+            "mining_dict": mining_dict,
+            "custom_payload": custom_payload,
+            "additional_args": additional_args
+        }
+        logger.info(f"🎯 Starting Dalfox XSS scan: {url if url else 'pipe mode'}")
+        result = hexstrike_client.safe_post("api/tools/dalfox", data)
+        if result.get("success"):
+            logger.info(f"✅ Dalfox XSS scan completed")
+        else:
+            logger.error(f"❌ Dalfox XSS scan failed")
+        return result
